Nunca mais trabalho com fragmentos
Porque os fragmentos parecem trabalhadores do estado e demoram até estarem funcionais, é preciso um check rápido se podemos aceder ao requireContext() e afins.
Está resolvido

+++ app/src/main/java/com/example/parkingsystem/api/matricula/MatriculaEndpoint.kt
@@ -12,7 +12,7 @@ import retrofit2.http.*
 interface MatriculaEndpoint {
 
     @GET("/api/matricula/getMatriculaUtilizador/{id}")
-    fun getMatriculaUtilizador(@Path("id") id: Int): Call<List<Matricula>>
+    fun getMatriculaUtilizador(@Path("id") id: Long): Call<List<Matricula>>
 
     @GET("/api/matricula/updateMatriculaSelected?idUtilizador&idMatricula")
     fun updateMatricula(@Query("idUtilizador") idUtilizador: Int, @Query("matricula") matricula: String): Call<List<Matricula>>
